# ECommerce.Api/Swaggerv1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "E-Mall API v1",
    "version": "v1"
  },
  "paths": {
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products from the database.",
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductViewModel"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Add a new product to the database.",
        "operationId": "Add",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "String",
              "default": "Bearer "
            }
          }
        ],
        "requestBody": {
          "description": "ProductViewModel Object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a specific product from the database.",
        "operationId": "Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Product ID",
              "format": "int64"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update an existing product in the database.",
        "operationId": "Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Product ID",
              "format": "int64"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "String",
              "default": "Bearer "
            }
          }
        ],
        "requestBody": {
          "description": "ProductViewModel Object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductViewModel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Products/{userId}/{id}": {
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Mark a product as not available instead of truly deleting it.",
        "operationId": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Product ID",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "nullable": true
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token",
            "required": true,
            "schema": {
              "type": "String",
              "default": "Bearer "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "required": [
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "maximum": 99999.99,
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProductViewModel": {
        "required": [
          "categoryIds"
        ],
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "categoryIds": {
            "minLength": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "additionalProperties": false
      }
    }
  }
}